Marco Clemencic  - Fixed few minor problems (including a possible buffer overflow).
git-svn-id: svn+ssh://svn.cern.ch/reps/gaudi/Gaudi/trunk@6027 53ecefc3-a64d-0410-9bdf-b8581fa3287e

+++ GaudiKernel/src/Lib/System.cpp
@@ -331,7 +331,7 @@ std::string __typeName(char*&  name) {
   else  {
     std::string result;
     char* ptr;
-    int i = ::strtol(name, &ptr, 10);
+    long i = ::strtol(name, &ptr, 10);
     if ( i <= 0 )  {
       name = ptr;
       while ( *name != 0 && *name != 'Z' )  {
